truba/requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
Django==1.10
django-admin-sortable2==0.6.10
django-angular==0.8.4
django-appconf==1.0.2
django-ckeditor==5.3.0
django-classy-tags==0.8.0
django-cms==3.3.3
django-filer==1.2.7
django-formtools==2.0
django-fsm==2.5.0
django-fsm-admin==1.2.4
django-ipware==1.1.6
django-js-asset==0.1.1
django-mptt==0.8.7
django-polymorphic==1.0.2
django-post-office==2.0.8
django-rest-auth==0.9.1
django-sekizai==0.10.0
Django-Select2==5.8.10
django-shop==0.9.3
django-suit==0.2.25
django-tinymce==2.6.0
django-treebeard==4.1.0
django-zurb-foundation==5.5.0
djangocms-admin-style==1.2.6.2
djangocms-cascade==0.12.0
djangorestframework==3.6.2
easy-thumbnails==2.3
enum34==1.1.6
freeze==1.0.10
idna==2.5
jsonfield==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
requests==2.18.1
setupfiles==0.0.50
six==1.10.0
Unidecode==0.4.20
urllib3==1.21.1
